@@ -2,13 +2,13 @@ name: flutter_chromecast_example
 version: 0.0.2
 description: An example implementation of the flutter_mdsn_plugin and the dart_chromecast package
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=3.6.0 <4.0.0'
 dependencies:
   flutter:
     sdk: flutter
   shared_preferences: ^2.0.11
   cupertino_icons: ^1.0.2
-  flutter_cast_video: ^1.2.0
+  cast: ^2.1.0
 
 dev_dependencies:
   flutter_test:
