@@ -1,5 +1,5 @@
+import 'package:cast/cast.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_cast_video/flutter_cast_video.dart';
 
 void main() {
   runApp(MyApp());
@@ -8,7 +8,21 @@ void main() {
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(home: CastSample());
+    return MaterialApp(
+      // Egyszerű karácsonyi színek használata
+      theme: ThemeData(
+        primarySwatch: Colors.red,
+        scaffoldBackgroundColor: Colors.green.shade100,
+        appBarTheme: AppBarTheme(color: Colors.redAccent),
+      ),
+      home: Scaffold(
+        appBar: AppBar(
+          title: Text('Karácsonyi Chromecast'),
+          centerTitle: true,
+        ),
+        body: CastSample(),
+      ),
+    );
   }
 }
 
@@ -24,149 +38,212 @@ class CastSample extends StatefulWidget {
 }
 
 class _CastSampleState extends State<CastSample> {
-  ChromeCastController _controller;
-  AppState _state = AppState.idle;
-  bool _playing = false;
+  Future<List<CastDevice>>? _future;
+  CastSession? _session; // A kiválasztott eszközzel létrehozott session.
+  bool _isConnecting = false; // Hogy tudjuk, éppen történik-e csatlakozás.
+  CastDevice? _selectedDevice;
 
-  var currentMp3 = CastSample.csengo;
   @override
   void initState() {
     super.initState();
+    _startSearch();
+  }
 
-    currentMp3 = CastSample.csengo;
+  void _startSearch() {
+    _future = CastDiscoveryService().search();
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Christmas Appc'),
-        actions: <Widget>[
-          ChromeCastButton(
-            size: CastSample._iconSize,
-            color: Colors.white,
-            onButtonCreated: _onButtonCreated,
-            onSessionStarted: _onSessionStarted,
-            onSessionEnded: () => setState(() => _state = AppState.idle),
-            onRequestCompleted: _onRequestCompleted,
-            onRequestFailed: _onRequestFailed,
+    // Ha már van session, jelenítsük meg a gombokat a lejátszáshoz.
+    if (_session != null) {
+      return _buildControlButtons();
+    }
+
+    // Különben listázzuk az eszközöket (vagy azt, hogy épp csatlakozás folyik).
+    return _buildDeviceList();
+  }
+
+  /// Eszközlista megjelenítése, ha még nincs csatlakozva
+  Widget _buildDeviceList() {
+    return FutureBuilder<List<CastDevice>>(
+      future: _future,
+      builder: (context, snapshot) {
+        if (snapshot.hasError) {
+          return Center(
+            child: Text(
+              'Hiba történt: ${snapshot.error.toString()}',
+              style: TextStyle(color: Colors.red),
+            ),
+          );
+        } else if (!snapshot.hasData) {
+          return Center(child: CircularProgressIndicator());
+        }
+
+        final devices = snapshot.data ?? [];
+        if (devices.isEmpty) {
+          return Center(child: Text('Nem található Chromecast eszköz.'));
+        }
+
+        return ListView.builder(
+          itemCount: devices.length,
+          itemBuilder: (context, index) {
+            final device = devices[index];
+            return ListTile(
+              title: Text(device.name),
+              trailing: _isConnecting
+                  ? SizedBox(
+                      width: 24, height: 24, child: CircularProgressIndicator())
+                  : Icon(Icons.cast),
+              onTap: () async {
+                // Ne lehessen többször rákattintani, amíg folyik a csatlakozás
+                if (_isConnecting) return;
+
+                setState(() {
+                  _isConnecting = true;
+                });
+
+                await _connectToDevice(device);
+
+                setState(() {
+                  _isConnecting = false;
+                });
+              },
+            );
+          },
+        );
+      },
+    );
+  }
+
+  /// Két gomb (csengő, mennyből az angyal) megjelenítése, ha már csatlakoztunk
+  Widget _buildControlButtons() {
+    return Padding(
+      padding: const EdgeInsets.all(16.0),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(
+            'Csatlakozva: ${_selectedDevice?.name ?? "Ismeretlen eszköz"}',
+            style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+          ),
+          SizedBox(height: 24),
+          ElevatedButton.icon(
+            icon: Icon(Icons.music_note),
+            label: Text("Csengő lejátszása"),
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Colors.redAccent,
+              padding: EdgeInsets.symmetric(vertical: 12, horizontal: 24),
+            ),
+            onPressed: () => _playMedia(CastSample.csengo, "Csengő"),
+          ),
+          SizedBox(height: 16),
+          ElevatedButton.icon(
+            icon: Icon(Icons.music_video),
+            label: Text("Mennyből az angyal lejátszása"),
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Colors.green,
+              padding: EdgeInsets.symmetric(vertical: 12, horizontal: 24),
+            ),
+            onPressed: () =>
+                _playMedia(CastSample.menybol, "Mennyből az angyal"),
+          ),
+          SizedBox(height: 16),
+          ElevatedButton.icon(
+            icon: Icon(Icons.stop),
+            label: Text("Lejátszás leállítása"),
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Colors.grey,
+              padding: EdgeInsets.symmetric(vertical: 12, horizontal: 24),
+            ),
+            onPressed: _stopPlayback,
           ),
         ],
       ),
-      body: Center(child: _handleState()),
     );
   }
 
-  Widget _handleState() {
-    switch (_state) {
-      case AppState.idle:
-        return Text('ChromeCast not connected');
-      case AppState.connected:
-        return Text('No media loaded');
-      case AppState.mediaLoaded:
-        return _mediaControls();
-      case AppState.error:
-        return Text('An error has occurred');
-      default:
-        return Container();
+  /// Csatlakozás a kiválasztott eszközhöz
+  Future<void> _connectToDevice(CastDevice device) async {
+    try {
+      final session = await CastSessionManager().startSession(device);
+
+      session.stateStream.listen((state) {
+        if (state == CastSessionState.connected) {
+          setState(() {
+            _selectedDevice = device;
+          });
+
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(content: Text('Sikeres csatlakozás: ${device.name}')),
+          );
+        } else if (state == CastSessionState.closed ||
+            state == CastSessionState.closed) {
+          // Ha bontja a kapcsolatot, akkor térjünk vissza az eszközlistához
+          setState(() {
+            _selectedDevice = null;
+            _session = null;
+          });
+        }
+      });
+
+      // Ha minden rendben, eltároljuk a session-t
+      setState(() {
+        _session = session;
+      });
+
+      // Indítjuk a "receiver" appot (pl. a default media receiver: CC1AD845)
+      session.sendMessage(CastSession.kNamespaceReceiver, {
+        'type': 'LAUNCH',
+        'appId': 'CC1AD845', // Alapértelmezett media receiver
+      });
+    } catch (e) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(content: Text('Hiba a csatlakozás során: $e')),
+      );
     }
   }
 
-  Widget _mediaControls() {
-    return Column(
-      children: [
-        RadioListTile(
-          title: Text("Csengő"),
-          value: CastSample.csengo,
-          groupValue: currentMp3,
-          onChanged: (value) async {
-            _controller.stop();
-            await _controller.loadMedia(CastSample.csengo);
-            setState(() {
-              currentMp3 = CastSample.csengo;
-            });
-            _playPauseMenybol();
-          },
-        ),
-        RadioListTile(
-          title: Text("Menyből az angyal"),
-          value: CastSample.menybol,
-          groupValue: currentMp3,
-          onChanged: (value) async {
-            _controller.stop();
-            await _controller.loadMedia(CastSample.menybol);
-            setState(() {
-              currentMp3 = CastSample.menybol;
-            });
-            _playPauseMenybol();
-          },
-        ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            _RoundIconButton(
-              icon: Icons.replay_10,
-              onPressed: () =>
-                  _controller.seek(relative: true, interval: -10.0),
-            ),
-            _RoundIconButton(
-                icon: _playing ? Icons.pause : Icons.play_arrow,
-                onPressed: _playPauseMenybol),
-            _RoundIconButton(
-              icon: Icons.forward_10,
-              onPressed: () => _controller.seek(relative: true, interval: 10.0),
-            )
-          ],
-        ),
-      ],
+  /// Leállít bármilyen aktív lejátszást
+  void _stopPlayback() {
+    if (_session == null) return;
+    _session!.sendMessage(
+      CastSession.kNamespaceMedia,
+      {
+        "type": "STOPg",
+        "sessionId": _session!.sessionId,
+      },
+    );
+    _session?.close();
+    print("Leállítás");
+  }
+
+  /// Egy általános függvény, ami lejátszik valamilyen médiaállományt
+  void _playMedia(String contentId, String title) {
+    if (_session == null) return;
+
+    final message = {
+      'contentId': contentId,
+      'contentType': 'audio/mp3', // Mindkét esetben mp3
+      'streamType': 'BUFFERED',
+      'metadata': {
+        'type': 0,
+        'metadataType': 0,
+        'title': title,
+        'images': [
+          {'url': contentId}
+        ],
+      }
+    };
+
+    _session!.sendMessage(
+      CastSession.kNamespaceMedia,
+      {
+        'type': 'LOAD',
+        'autoPlay': true,
+        'currentTime': 0,
+        'media': message,
+      },
     );
   }
-
-  Future<void> _playPauseMenybol() async {
-    final playing = await _controller.isPlaying();
-    if (playing) {
-      await _controller.pause();
-    } else {
-      await _controller.play();
-    }
-    setState(() => _playing = !playing);
-  }
-
-  Future<void> _onButtonCreated(ChromeCastController controller) async {
-    _controller = controller;
-    await _controller.addSessionListener();
-  }
-
-  Future<void> _onSessionStarted() async {
-    setState(() => _state = AppState.connected);
-    await _controller.loadMedia(CastSample.csengo);
-  }
-
-  Future<void> _onRequestCompleted() async {
-    final playing = await _controller.isPlaying();
-    setState(() {
-      _state = AppState.mediaLoaded;
-      _playing = playing;
-    });
-  }
-
-  Future<void> _onRequestFailed(String error) async {
-    setState(() => _state = AppState.error);
-    print(error);
-  }
 }
-
-class _RoundIconButton extends StatelessWidget {
-  final IconData icon;
-  final VoidCallback onPressed;
-
-  _RoundIconButton({@required this.icon, @required this.onPressed});
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-        child: Icon(icon, color: Colors.white), onPressed: onPressed);
-  }
-}
-
-enum AppState { idle, connected, mediaLoaded, error }
